projects/api_gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = ""
authors = ["pai-cm <cmkim@publicai.co.kr>"]
readme = "README.md"
packages = [
    {include="src"}
]

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.20.0"
pyjwt = "^2.8.0"
cryptography = "^42.0.5"
fastapi = "^0.110.1"
uvicorn = "^0.29.0"
httpx = "^0.27.0"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
